refactor: drop `weaver note`, add a schema migration framework
`weaver note` duplicated the `set-status` modality. The status-description
trail — each `set-status` message, recorded as an `attention` event — already
captures progress, so a separate free-form note log was redundant.

Remove the note feature end to end:
- the `weaver note` command and the `note` weaver-core module
- the `notes` table, the `POST /api/sessions/{id}/note` route + handler, and
  loom's note-on-merge write
- the `note` event kind from the UI activity feeds (the status message lives
  on as the `attention` event's `note` field / the branch description)

`weaver summary` now surfaces the latest status message ("last status") drawn
from the status-description history instead of the last note.

Introduce a real migration framework (`weaver_core::migrations`): ordered SQL
files under `crates/weaver-core/migrations/` (embedded with `include_str!`),
applied at startup and recorded in a `schema_migrations` indicator table so
each runs exactly once. A one-time `legacy_bootstrap` brings pre-framework
databases (the old `workspaces` drop, the issues repo-scope reshape, additive
columns) up to the baseline on first run. Migration 0002 drops the `notes`
table.
