+++ crates/loom/tests/integration/archive.rs
@@ -40,16 +40,10 @@ async fn archive_keeps_branch_and_history() {
         "archive worktree missing"
     );
 
-    // A note proves the weaver history survives the archive.
-    client
-        .post(
-            &format!("/api/sessions/{arch_id}/note"),
-            json!({ "text": "decision: keep going" }),
-        )
-        .await
-        .unwrap();
     // Flag the session for attention; archiving must clear it (a torn-down
-    // workstream can't still "need me").
+    // workstream can't still "need me"). The recorded `attention` event (with
+    // its `source: manual` marker) doubles as branch history we expect to
+    // survive the archive.
     client
         .patch(
             &format!("/api/sessions/{arch_id}"),
@@ -89,14 +83,15 @@ async fn archive_keeps_branch_and_history() {
         weaver_core::git::branch_exists(ts.repo_path(), "weaver/archive-me").await,
         "archive must not delete the branch"
     );
-    // The note history survives in the branch log.
+    // The branch event history survives the archive (unlike delete): the
+    // pre-archive manual attention event is still in the log.
     let log = client
         .get(&format!("/api/sessions/{arch_id}/log"))
         .await
         .unwrap();
     assert!(
-        serde_json::to_string(&log).unwrap().contains("keep going"),
-        "note history should survive archive"
+        serde_json::to_string(&log).unwrap().contains("manual"),
+        "branch history should survive archive"
     );
 
     // An archived session can still be fully removed afterwards.
