+++ crates/weaver-core/src/agent.rs
@@ -81,7 +81,7 @@ mod tests {
         assert!(v["hookSpecificOutput"]["additionalContext"]
             .as_str()
             .unwrap()
-            .contains("weaver note"));
+            .contains("weaver set-status"));
     }
 
     #[test]
