+++ crates/weaver/tests/agent_cli.rs
@@ -280,8 +280,9 @@ fn summary_orients_an_agent_on_the_branch() {
     run(&env, &["goal", "ship", "the", "feature"]);
     run(&env, &["issue", "add", "wire", "up", "routes"]);
     run(&env, &["issue", "add", "add", "tests"]);
+    // Two status updates: the latest is what "last status" surfaces.
+    run(&env, &["set-status", "blocked", "build", "broken"]);
     run(&env, &["set-status", "ok", "routes", "wired"]);
-    run(&env, &["note", "left", "off", "mid-refactor"]);
 
     let out = run(&env, &["summary"]);
     assert!(out.contains("ship the feature"), "summary: {out}");
@@ -290,11 +291,9 @@ fn summary_orients_an_agent_on_the_branch() {
     assert!(out.contains("Outstanding (2):"), "summary: {out}");
     assert!(out.contains("#1    wire up routes"), "summary: {out}");
     assert!(out.contains("#2    add tests"), "summary: {out}");
-    // Hints: the most recent note plus a next-action pointing at the first task.
-    assert!(
-        out.contains("last note: left off mid-refactor"),
-        "summary: {out}"
-    );
+    // Hints: the latest status message (from the status-description trail) plus
+    // a next-action pointing at the first task.
+    assert!(out.contains("last status: routes wired"), "summary: {out}");
     assert!(out.contains("pick up #1"), "summary: {out}");
     // Every section advertises the command that drills into it.
     for hint in [
@@ -340,14 +339,6 @@ fn summary_with_no_open_tasks_suggests_wrapping_up() {
     assert!(out.contains("Outstanding: none"), "summary: {out}");
     assert!(out.contains("no open tasks"), "summary: {out}");
     assert!(out.contains("open a PR"), "summary: {out}");
-}
-
-#[test]
-fn note_writes_an_event() {
-    let env = setup();
-    run(&env, &["note", "made", "a", "decision"]);
-    let log = run(&env, &["log"]);
-    assert!(log.contains("made a decision"), "log: {log}");
 }
 
 #[test]
